rename sample base file
